requirements.txt
streamlit
openai-whisper
torch
transformers
ffmpeg-python
numpy
regex
pydub